// pc/src/services/apis/models/ProjectDcdxCompletionInvestmentVo.ts
export interface ProjectDcdxCompletionInvestmentVo {
    /**
     * 主键
     * @type {string}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    id?: string;
    /**
     * 市区名称
     * @type {string}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    district?: string;
    /**
     * 园区名称
     * @type {string}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    park?: string;
    /**
     * 年份
     * @type {number}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    year?: number;
    /**
     * 竣工项目数
     * @type {number}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    completedProjectCount?: number;
    /**
     * 签约投资额（万元）
     * @type {number}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    signedInvestment?: number;
    /**
     * 完成投资额（万元）
     * @type {number}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    completedInvestment?: number;
    /**
     * 其中：固定资产投资（万元）
     * @type {number}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    fixedAssetInvestment?: number;
    /**
     * 设备投资（万元）
     * @type {number}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    equipmentInvestment?: number;
    /**
     * 投资完成比重（完成/签约，小数形式）
     * @type {number}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    investmentCompletionRatio?: number;
    /**
     * 所属产业链群
     * @type {string}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    industrialChainCluster?: string;
    /**
     * 是否产业链群
     * @type {boolean}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    isLq?: boolean;
    /**
     * 
     * @type {string}
     * @memberof ProjectDcdxCompletionInvestmentVo
     */
    children: string;
}